backend/requirements.txt
# ==========================================
# CodeGuard AI - Backend Dependencies
# ==========================================

# ===== WEB FRAMEWORK =====
fastapi>=0.109.0  # Última versión compatible con Pydantic 2.8+
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# ===== DATABASE & ORM =====
sqlalchemy>=2.0.25
alembic>=1.13.0
psycopg2-binary>=2.9.9

# ===== AUTHENTICATION =====
# Clerk (requiere Pydantic 2.8+, httpx 0.27+)
clerk-backend-api>=1.0.0
python-jose[cryptography]>=3.3.0

# ===== SUPABASE (comentado - incompatible con Clerk httpx) =====
# Usar PostgreSQL directo con SQLAlchemy en su lugar
# supabase>=2.3.0  

# ===== STATIC ANALYSIS (Sprint 1 Core) =====
bandit>=1.7.5
radon>=6.0.1
pylint>=3.0.3
flake8>=7.0.0

# ===== CACHE =====
redis>=5.0.1
hiredis>=2.3.0

# ===== UTILITIES =====
pydantic>=2.8.0  # Compatible con Clerk
pydantic-settings>=2.2.0
python-dotenv>=1.0.0
requests>=2.31.0
httpx>=0.27.0  # Compatible con Clerk

# ===== AI SERVICES (Sprint 3) =====
google-generativeai>=0.3.2  # Gemini API
google-cloud-aiplatform>=1.40.0  # Vertex AI (opcional)

# ===== DEVELOPMENT TOOLS =====  (Moved to requirements-dev.txt)
# black>=24.0.0
# isort>=5.13.0
# mypy>=1.8.0

# ===== TESTING (Sprint 1 Core) =====  (Moved to requirements-dev.txt)
# pytest>=8.0.0
# pytest-asyncio>=0.23.0
# pytest-cov>=4.1.0
# faker>=22.0.0  # Para datos de prueba

# ===== PDF GENERATION (Sprint 2) =====
reportlab>=4.0.9

# ===== WEBSOCKETS (Sprint 2) =====
websockets>=12.0
email-validator>=2.0.0
